# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Inventario",
        "description": "Documentación de la API del sistema de inventario",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/suppliers": {
            "get": {
                "tags": [
                    "Proveedores"
                ],
                "summary": "Obtener todos los proveedores.",
                "operationId": "6207e57dd1dcd1d6e7125fe144607196",
                "responses": {
                    "200": {
                        "description": "Lista de proveedores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No hay datos, por favor registre al menos un proveedor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Proveedores"
                ],
                "summary": "Crear un nuevo proveedor. Requiere tener el rol Admin.",
                "operationId": "dc5dc7ccb140cd5b81cab6f42b396691",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone",
                                    "address",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Proveedor S.A."
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "proveedor@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Calle 123, Ciudad"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Proveedor mayorista de productos"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proveedor creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Proveedor creado exitosamente"
                                        },
                                        "suppliers": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/suppliers/{id}": {
            "get": {
                "tags": [
                    "Proveedores"
                ],
                "summary": "Obtener un proveedor por ID.",
                "operationId": "7fc447855461518ce36bdeb5a44310a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Proveedor encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Proveedor encontrado"
                                        },
                                        "supplier": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Proveedor no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Proveedores"
                ],
                "summary": "Actualizar un proveedor existente. Requiere tener el rol Admin.",
                "operationId": "22145931fdda85497cf90c73c5406747",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone",
                                    "address",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Proveedor S.A. Actualizado"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "nuevo@email.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "987654321"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Av. Siempre Viva 742"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción actualizada del proveedor"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proveedor actualizado de manera exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Proveedor actualizado de manera exitosa"
                                        },
                                        "category": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Proveedor no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Proveedores"
                ],
                "summary": "Eliminar un proveedor existente. Requiere tener el rol Admin.",
                "operationId": "2fe04361f0f1581a9f5dbb4094b634b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proveedor eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Proveedor eliminado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Proveedor no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}